+++ src/bpp/models/cache.py
@@ -190,17 +190,32 @@ class RekordManager(FulltextSearchMixin, models.Manager):
             .distinct()
         )
 
-    def prace_jednostki(self, jednostka):
+    def prace_jednostki(self, jednostka, afiliowane=None):
+        if afiliowane is None:
+            return self.filter(
+                autorzy__jednostka__pk__in=list(
+                    jednostka.get_descendants(include_self=True).values_list(
+                        "pk", flat=True
+                    )
+                )
+            ).distinct()
+
         return self.filter(
             autorzy__jednostka__pk__in=list(
                 jednostka.get_descendants(include_self=True).values_list(
                     "pk", flat=True
                 )
-            )
+            ),
+            autorzy__afiliuje=afiliowane,
         ).distinct()
 
-    def prace_wydzialu(self, wydzial):
-        return self.filter(autorzy__jednostka__wydzial=wydzial).distinct()
+    def prace_wydzialu(self, wydzial, afiliowane=None):
+        if afiliowane is None:
+            return self.filter(autorzy__jednostka__wydzial=wydzial).distinct()
+
+        return self.filter(
+            autorzy__jednostka__wydzial=wydzial, autorzy__afiliuje=afiliowane
+        ).distinct()
 
     def redaktorzy_z_jednostki(self, jednostka):
         return self.filter(